Remove ManagedEntity subscript for non-optional KeyPaths.
Since a ManagedEntity always wraps an optional value,
building a new ManagedEntity from a WritableKeyPath that
leads to a non-optional Entity is problematic, since it
is unclear what writing a nil value to that would
entail. As I am not using this subscript anywhere in the
code, I am just removing it for now.