Update the code to make it build
The builds were failing, because there were
declarations in conditional statements, that had
right-hand sides that weren't optional. It clearly
used to build that way, but now it doesn't. I can't
find a mention of this behavior being changed anywhere
and I am not going to downgrade XCode to chase down
which Version exactly broke this. My best guess is,
that declarations in conditionals, that were not the
first condition in the conditional weren't required to
have optional right-hand sides previously and that
that was a bug and has been fixed in some
minor-version update of Swift.